fix: correct configration/exsist typos across codebase
Rename configration -> configuration (12 files): 2 file renames, all references in Rust/TS/i18n.

Rename exsist -> exists in groups.rs and project.rs.
