+++ src/locales/en.json
@@ -113,17 +113,17 @@
   "Mirror-Tip": "Please select or input the appropriate mirror url according to your region.",
   "Default": "Default",
   "Custom": "Custom",
-  "Configration": "Configration",
-  "Configration-export": "Configration export",
-  "Configration-export-tip": "Please select the configuration you want to export:",
-  "Configration-export-setting": "(lanuage, installation directory, mirror url, etc.)",
-  "Configration-export-projects": "(include groups)",
-  "Configration-export-success": "The export has been completed, please view it in the {{filename}} file",
-  "Configration-import": "Configration import",
-  "Configration-import-tip": "Whether to automatically synchronize the nodejs version (.nvmdrc) for project data when importing it?",
+  "Configuration": "Configuration",
+  "Configuration-export": "Configuration export",
+  "Configuration-export-tip": "Please select the configuration you want to export:",
+  "Configuration-export-setting": "(language, installation directory, mirror url, etc.)",
+  "Configuration-export-projects": "(include groups)",
+  "Configuration-export-success": "The export has been completed, please view it in the {{filename}} file",
+  "Configuration-import": "Configuration import",
+  "Configuration-import-tip": "Whether to automatically synchronize the nodejs version (.nvmdrc) for project data when importing it?",
   "Import-only": "Import only",
   "Import-and-sync": "Import and sync",
-  "Configration-import-success": "The import has been completed. If you encounter data inconsistency, please restart the application.",
+  "Configuration-import-success": "The import has been completed. If you encounter data inconsistency, please restart the application.",
   "Groups": "Groups",
   "Group-Name": "Group Name",
   "Group-Desc": "Group Description",
