+++ src/locales/pl.json
@@ -113,17 +113,17 @@
   "Mirror-Tip": "Wybierz lub wpisz odpowiedni adres mirror dla swojego regionu.",
   "Default": "Domyślny",
   "Custom": "Niestandardowy",
-  "Configration": "Konfiguracja",
-  "Configration-export": "Eksport konfiguracji",
-  "Configration-export-tip": "Wybierz konfigurację do eksportu:",
-  "Configration-export-setting": "(język, katalog instalacji, mirror itd.)",
-  "Configration-export-projects": "(wraz z grupami)",
-  "Configration-export-success": "Eksport zakończony, sprawdź plik {{filename}}",
-  "Configration-import": "Import konfiguracji",
-  "Configration-import-tip": "Czy automatycznie synchronizować wersję nodejs (.nvmdrc) dla danych projektu podczas importu?",
+  "Configuration": "Konfiguracja",
+  "Configuration-export": "Eksport konfiguracji",
+  "Configuration-export-tip": "Wybierz konfigurację do eksportu:",
+  "Configuration-export-setting": "(język, katalog instalacji, mirror itd.)",
+  "Configuration-export-projects": "(wraz z grupami)",
+  "Configuration-export-success": "Eksport zakończony, sprawdź plik {{filename}}",
+  "Configuration-import": "Import konfiguracji",
+  "Configuration-import-tip": "Czy automatycznie synchronizować wersję nodejs (.nvmdrc) dla danych projektu podczas importu?",
   "Import-only": "Tylko import",
   "Import-and-sync": "Import i synchronizacja",
-  "Configration-import-success": "Import zakończony. W przypadku niespójności danych uruchom ponownie aplikację.",
+  "Configuration-import-success": "Import zakończony. W przypadku niespójności danych uruchom ponownie aplikację.",
   "Groups": "Grupy",
   "Group-Name": "Nazwa grupy",
   "Group-Desc": "Opis grupy",
