+++ src/locales/zh_CN.json
@@ -113,17 +113,17 @@
   "Mirror-Tip": "请根据您所在的地区选择或者输入适合的镜像地址。",
   "Default": "默认",
   "Custom": "自定义",
-  "Configration": "配置",
-  "Configration-export": "配置导出",
-  "Configration-export-tip": "请选择您要导出的配置：",
-  "Configration-export-setting": "（语言、安装目录、镜像地址等。）",
-  "Configration-export-projects": "（包括分组）",
-  "Configration-export-success": "导出已完成，请在 {{filename}} 文件中查看",
-  "Configration-import": "配置导入",
-  "Configration-import-tip": "导入项目数据时是否自动同步 nodejs 版本（.nvmdrc）？",
+  "Configuration": "配置",
+  "Configuration-export": "配置导出",
+  "Configuration-export-tip": "请选择您要导出的配置：",
+  "Configuration-export-setting": "（语言、安装目录、镜像地址等。）",
+  "Configuration-export-projects": "（包括分组）",
+  "Configuration-export-success": "导出已完成，请在 {{filename}} 文件中查看",
+  "Configuration-import": "配置导入",
+  "Configuration-import-tip": "导入项目数据时是否自动同步 nodejs 版本（.nvmdrc）？",
   "Import-only": "仅导入",
   "Import-and-sync": "导入且同步",
-  "Configration-import-success": "导入已完成。 如果遇到数据不一致，请重启应用。",
+  "Configuration-import-success": "导入已完成。 如果遇到数据不一致，请重启应用。",
   "Groups": "分组",
   "Group-Name": "分组名称",
   "Group-Desc": "分组描述",
